# spec/openapi-plugin.yaml
openapi: 3.1.0
info:
  title: Vyshka Plugin API
  version: 0.4.0
  summary: Game-server-facing API of the Vyshka hub.
  description: |
    The plugin half of the Vyshka protocol. Normative text lives in
    `spec/protocol.md`; this document is its machine-readable companion and
    covers only the endpoints implemented so far (spec sections 3.1.2, 5.2 and
    5.3).

    Three credentials, each with one job: a one-time enrollment token bootstraps
    a record into a permanent server secret, and the secret buys short-lived
    session tokens. Enroll and session authenticate through the request body,
    because they are what a plugin has before it has a bearer token; every other
    Plugin API call uses `Authorization: Bearer <sessionToken>`.

    Once a session exists, the plugin lives in `POST /plugin/v1/poll`: it keeps
    one request in flight at all times, and the hub answers it the moment there
    is anything to say.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://127.0.0.1:8080
    description: A hub running locally with default settings.
...
paths:
  /plugin/v1/enroll:
    parameters:
      - $ref: "#/components/parameters/InlineErrors"
    post:
      tags: [enrollment]
      operationId: enroll
      summary: Exchange a one-time enrollment token for server credentials.
      description: |
        The token is burned atomically: two plugins racing on the same token
        produce exactly one success. The secret is returned here and nowhere
        else. Enrolling over existing credentials (with a freshly issued token)
        replaces the secret and ends every session derived from the old one.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrollmentToken, game]
              properties:
                enrollmentToken:
                  type: string
                  description: The one-time token from the operator, opaque.
                game:
                  type: string
                  maxLength: 64
                  description: Game this plugin serves, lowercase; compared case-insensitively.
                  examples: [dayz, reforger]
                plugin: { $ref: "#/components/schemas/PluginDescriptor" }
                transports:
                  type: array
                  items: { type: string, enum: [poll, ws] }
                  description: Transports this plugin can speak. `poll` is mandatory.
            example:
              enrollmentToken: vye_...
              game: dayz
              plugin: { name: vyshka-dayz, version: 1.4.0 }
              transports: [poll]
      responses:
        "201":
          description: Permanent credentials for this server.
          content:
            application/json:
              schema:
                type: object
                required: [serverId, serverSecret, server]
                properties:
                  serverId: { type: string }
                  serverSecret:
                    type: string
                    description: Persist this. It is not retrievable again.
                  server: { $ref: "#/components/schemas/ServerIdentity" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: |
            The enrollment token is unknown or expired (`enrollment_token_invalid`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: |
            The token was already burned (`enrollment_token_used`), or the server
            record is registered for a different game (`game_mismatch`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /plugin/v1/session:
    parameters:
      - $ref: "#/components/parameters/InlineErrors"
    post:
      tags: [session]
      operationId: createSession
      summary: Exchange server credentials for a short-lived session.
      description: |
        The response is authoritative for everything the plugin requested. A hub
        holds at most one live session per server: issuing a session invalidates
        the previous one.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serverId, serverSecret]
              properties:
                serverId: { type: string }
                serverSecret: { type: string }
                protocolVersion:
                  type: integer
                  description: Version the plugin speaks; omit for the hub's current one.
                pollTimeoutSeconds:
                  type: integer
                  minimum: 5
                  maximum: 60
                  default: 25
                  description: |
                    Requested hold time for long-poll requests. Honored within 5
                    to 60 seconds and clamped outside it (spec section 3.1.1).
                    The plugin must set its own HTTP response timeout to the
                    effective value plus 5 seconds.
                plugin: { $ref: "#/components/schemas/PluginDescriptor" }
                transports:
                  type: array
                  items: { type: string, enum: [poll, ws] }
      responses:
        "200":
          description: A live session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400":
          description: |
            Missing credentials (`bad_request`), or a protocol version this hub
            does not speak (`protocol_version_unsupported`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: |
            Credentials do not match (`credentials_invalid`) or were revoked
            (`credentials_revoked`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [session]
      operationId: getSession
      summary: Report the session behind a session token.
      description: |
        Lets a plugin confirm its session is still live without waiting for a
        poll to fail. The session token is never echoed back.
      security:
        - sessionToken: []
      responses:
        "200":
          description: The live session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/SessionInvalid" }
  /plugin/v1/poll:
    parameters:
      - $ref: "#/components/parameters/InlineErrors"
    post:
      tags: [transport]
      operationId: poll
      summary: Exchange envelopes with the hub over a held request.
      description: |
        One request carries both directions. The plugin reports the highest
        contiguous hub -> plugin `seq` it has durably processed and sends
        whatever it has; the hub applies both, then answers with everything
        queued for the plugin.

        The hub holds the request up to the session's `pollTimeoutSeconds` and
        answers early as soon as anything is queued. An empty batch is the
        normal answer to a hold that expired with nothing to send, and the
        plugin re-polls immediately either way.

        A plugin must set its own HTTP response timeout to the effective
        `pollTimeoutSeconds` plus 5 seconds (spec section 3.1.1). An aborted
        poll loses nothing: sequence numbers and acks make the next one recover
        whatever was undelivered.
      security:
        - sessionToken: []
      requestBody:
        required: false
        description: An empty body, or `{}`, is a valid idle poll.
        content:
          application/json:
            schema:
              type: object
              properties:
                ack:
                  type: integer
                  minimum: 0
                  default: 0
                  description: |
                    Highest contiguous hub -> plugin `seq` durably processed.
                    Cumulative: acking N acks everything at or below N. 0 acks
                    nothing.
                envelopes:
                  type: array
                  maxItems: 500
                  description: |
                    Envelopes the plugin is sending, in ascending `seq` order. A
                    hub accepts at least 200 per request.
                  items: { $ref: "#/components/schemas/PluginEnvelope" }
            example:
              ack: 4181
              envelopes:
                - v: 1
                  id: 01J5QN...
                  type: event.batch
                  seq: 918
                  ts: "2026-08-16T18:00:00Z"
                  body:
                    events:
                      - t: core.player.death
                        ts: "2026-08-16T17:59:58Z"
                        data: { weapon: M4A1, distance: 312.5 }
                      - t: example-mod.raid.started
                        data: { territoryId: t-19, attackers: 4 }
      responses:
        "200":
          description: |
            Envelopes for the plugin, possibly none. Answered no later than the
            session's `pollTimeoutSeconds`.
          content:
            application/json:
              schema:
                type: object
                required: [envelopes, ack, pollTimeoutSeconds, sessionExpiresAt]
                properties:
                  envelopes:
                    type: array
                    description: |
                      Ascending by `seq`. Always present; an empty batch is `[]`,
                      never `null`. Anything above the plugin's ack is repeated
                      here unchanged until it is acked.
                    items: { $ref: "#/components/schemas/HubEnvelope" }
                  ack:
                    type: integer
                    minimum: 0
                    description: |
                      Highest contiguous plugin -> hub `seq` the hub has durably
                      processed. The plugin drops its buffer up to this.
                  pollTimeoutSeconds:
                    type: integer
                    minimum: 5
                    maximum: 60
                  sessionExpiresAt: { type: string, format: date-time }
        "400":
          description: |
            An inbound envelope is missing `id`, `type` or `seq`, carries a
            `seq` below 1, exceeds a length limit, or declares an unsupported
            envelope version, including an explicit `0` (`envelope_invalid`,
            with `details.index` naming the offending envelope); `ack` is above
            the highest `seq` the hub has sent (`ack_out_of_range`); or the
            batch is too large (`bad_request`). A rejected poll applies nothing
            at all, including its `ack`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/SessionInvalid" }
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: The session token from `POST /plugin/v1/session`.
  schemas:
...
    PluginEnvelope:
      description: |
        What a plugin must emit. `ts` is required of the sender even though a
        receiver must not reject an envelope over it; `v` stays optional.
      allOf: [{ $ref: "#/components/schemas/Envelope" }]
      required: [id, type, seq, ts]
    HubEnvelope:
      description: |
        What a hub must emit. A hub stamps `v` on everything it sends, so
        unlike a plugin it may not omit it.
      allOf: [{ $ref: "#/components/schemas/Envelope" }]
      required: [id, type, seq, ts, v]
    PluginDescriptor:
      type: object
      description: What the plugin says about itself. Advisory; the hub never routes on it.
      properties:
        name: { type: string }
        version: { type: string }
    ServerIdentity:
      type: object
      description: The server record this plugin is attached to.
      required: [id, name, game]
      properties:
        id: { type: string }
        name: { type: string }
        game: { type: string }
    Session:
      type: object
      required:
        [sessionId, expiresAt, protocolVersion, envelopeVersion, pollTimeoutSeconds, transports, features, server]
      properties:
        sessionId: { type: string }
        sessionToken:
          type: string
          description: Present only in the response that mints the session.
        expiresAt:
          type: string
          format: date-time
          description: Request a new session before this; an expired token is `session_invalid`.
        protocolVersion: { type: integer }
        envelopeVersion:
          type: integer
          description: The `v` this hub stamps on envelopes (spec section 4).
        pollTimeoutSeconds:
          type: integer
          minimum: 5
          maximum: 60
          description: Effective hold time, authoritative for this session.
        transports:
          type: array
          items: { type: string }
          description: Transports the hub offers. May grow without a version bump.
        features:
          type: object
          additionalProperties: true
          description: Hub-declared flags. Plugins must tolerate unrecognized keys.
        server: { $ref: "#/components/schemas/ServerIdentity" }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier; branch on this.
            status:
              type: integer
              minimum: 400
              maximum: 599
              description: |
                The HTTP status the hub would have sent. Present only when the
                error travels inline in a 200 because the request carried
                `?errors=inline` (spec section 2.3); the fallback for a `code`
                the client does not recognize.
            message:
              type: string
              description: Human-readable. Never parse it.
            details:
              type: object
              additionalProperties: true
  parameters:
    InlineErrors:
      name: errors
      in: query
      required: false
      schema:
        type: string
        enum: [inline]
      description: |
        Ask for this request's refusals inline (spec section 2.3): every
        response the hub would have sent as a 4xx or 5xx arrives instead as a
        200 whose body is the Error schema with `error.status` filled in.
        Successes are unchanged, and a success body never carries a top-level
        `error` member. For engines whose HTTP client hides the status and
        body of a non-2xx response. Any value other than `inline` is refused
        with an ordinary 400. A hub that predates the option ignores it.
...
  responses:
    BadRequest:
      description: Malformed body, or a field missing, wrongly typed, or out of range.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    SessionInvalid:
      description: |
        The session token is expired, unknown, or superseded (`session_invalid`).
        Request a new session; do not re-enroll.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }